remove general fetch, convert to specific GithubApi fetch hook
