+++ plugins/scaffolder-frontend-auth/src/index.ts
@@ -1,2 +1,2 @@
 export { useAuth, type RequestUserCredentials } from './hooks/useAuth';
-export { useFetchWithAuth } from './hooks/useFetchWithAuth';
+export { useGithubApi } from './hooks/useGithubApi';
